Conflicts:
	Sources/Internal/Render/2D/FTFont.cpp
	Sources/Internal/Render/2D/TextBlock.cpp

Former-commit-id: 8fc2a8ee38177cd58660eeec2c4d5f8f50026264

+++ Sources/Internal/Render/2D/TextBlock.cpp
@@ -724,7 +724,7 @@ void TextBlock::CalculateCacheParams()
         textLayout.Seek(0);
         textLayout.FillList(lines, drawSize.dx, isMultilineBySymbolEnabled);
         fontHeight = font->GetFontHeight() + yOffset;
-        textMetrics.height = textMetrics.drawRect.dy = fontHeight * (int32)lines.size() - yOffset;
+        textMetrics.height = textMetrics.drawRect.dy = fontHeight * int32(lines.size()) - yOffset;
 
         DVASSERT_MSG(!lines.empty(), "Empty lines information");
 
@@ -755,7 +755,7 @@ void TextBlock::CalculateCacheParams()
                             textLayout.FillList(lines, drawSize.dx, isMultilineBySymbolEnabled);
                             charactersSizes = textLayout.GetCharSizes();
                             fontHeight = font->GetFontHeight() + yOffset;
-                            textMetrics.height = textMetrics.drawRect.dy = fontHeight * (int32)lines.size() - yOffset;
+                            textMetrics.height = textMetrics.drawRect.dy = fontHeight * int32(lines.size()) - yOffset;
                             break;
                         }
                     }
@@ -818,14 +818,14 @@ void TextBlock::CalculateCacheParams()
                 textLayout.FillList(lines, drawSize.dx, isMultilineBySymbolEnabled);
                 charactersSizes = textLayout.GetCharSizes();
                 fontHeight = font->GetFontHeight() + yOffset;
-                textMetrics.height = textMetrics.drawRect.dy = fontHeight * (int32)lines.size() - yOffset;
+                textMetrics.height = textMetrics.drawRect.dy = fontHeight * int32(lines.size()) - yOffset;
             };
         }
 
         // Detect visible lines
         if (textMetrics.height > drawSize.y && requestedSize.y >= 0.f)
         {
-            int32 needLines = Min((int32)lines.size(), (int32)ceilf(drawSize.y / fontHeight) + 1);
+            int32 needLines = Min(int32(lines.size()), int32(ceilf(drawSize.y / fontHeight)) + 1);
             Vector<TextLayout::Line> oldLines;
             lines.swap(oldLines);
             if (align & ALIGN_TOP)
@@ -834,15 +834,15 @@ void TextBlock::CalculateCacheParams()
             }
             else if (align & ALIGN_VCENTER)
             {
-                int32 startIndex = ((int32)oldLines.size() - needLines + 1) / 2;
+                int32 startIndex = (int32(oldLines.size()) - needLines + 1) / 2;
                 lines.assign(oldLines.begin() + startIndex, oldLines.begin() + startIndex + needLines);
             }
             else //if(ALIGN_BOTTOM)
             {
-                int32 startIndex = (int32)oldLines.size() - needLines;
+                int32 startIndex = int32(oldLines.size()) - needLines;
                 lines.assign(oldLines.begin() + startIndex, oldLines.end());
             }
-            textMetrics.height = textMetrics.drawRect.dy = fontHeight * (int32)lines.size() - yOffset;
+            textMetrics.height = textMetrics.drawRect.dy = fontHeight * int32(lines.size()) - yOffset;
         }
 
         // Get lines as visual strings and its metrics
@@ -901,7 +901,7 @@ void TextBlock::CalculateCacheParams()
 
     if (requestedSize.dx >= 0 && useJustify)
     {
-        textMetrics.drawRect.dx = Max(textMetrics.drawRect.dx, (int)drawSize.dx);
+        textMetrics.drawRect.dx = Max(textMetrics.drawRect.dx, int32(drawSize.dx));
     }
 
     float32 lyoffset = 0.f;
@@ -937,10 +937,10 @@ void TextBlock::CalculateCacheParams()
     }
 
     //calculate texture size
-    int32 dx = (int32)ceilf(VirtualCoordinatesSystem::Instance()->ConvertVirtualToPhysicalX((float32)textMetrics.drawRect.dx));
-    int32 dy = (int32)ceilf(VirtualCoordinatesSystem::Instance()->ConvertVirtualToPhysicalY((float32)textMetrics.drawRect.dy));
-    int32 ox = (int32)ceilf(VirtualCoordinatesSystem::Instance()->ConvertVirtualToPhysicalX((float32)textMetrics.drawRect.x));
-    int32 oy = (int32)ceilf(VirtualCoordinatesSystem::Instance()->ConvertVirtualToPhysicalY((float32)textMetrics.drawRect.y));
+    int32 dx = int32(ceilf(VirtualCoordinatesSystem::Instance()->ConvertVirtualToPhysicalX(float32(textMetrics.drawRect.dx))));
+    int32 dy = int32(ceilf(VirtualCoordinatesSystem::Instance()->ConvertVirtualToPhysicalY(float32(textMetrics.drawRect.dy))));
+    int32 ox = int32(ceilf(VirtualCoordinatesSystem::Instance()->ConvertVirtualToPhysicalX(float32(textMetrics.drawRect.x))));
+    int32 oy = int32(ceilf(VirtualCoordinatesSystem::Instance()->ConvertVirtualToPhysicalY(float32(textMetrics.drawRect.y))));
 
     cacheUseJustify = useJustify;
     cacheDx = dx;
@@ -953,34 +953,34 @@ void TextBlock::CalculateCacheParams()
     cacheOy = oy;
 
     cacheW = textMetrics.drawRect.dx;
-    cacheFinalSize.x = (float32)textMetrics.drawRect.dx;
-    cacheFinalSize.y = (float32)textMetrics.drawRect.dy;
-    cacheTextSize = Vector2((float32)textMetrics.width, (float32)textMetrics.height);
+    cacheFinalSize.x = float32(textMetrics.drawRect.dx);
+    cacheFinalSize.y = float32(textMetrics.drawRect.dy);
+    cacheTextSize = Vector2(float32(textMetrics.width), float32(textMetrics.height));
 
     // Align sprite offset
     if (align & ALIGN_RIGHT)
     {
-        cacheSpriteOffset.x = (float32)(textMetrics.drawRect.dx - textMetrics.width + textMetrics.drawRect.x);
+        cacheSpriteOffset.x = float32(textMetrics.drawRect.dx - textMetrics.width + textMetrics.drawRect.x);
     }
     else if (align & ALIGN_HCENTER)
     {
-        cacheSpriteOffset.x = ((float32)(textMetrics.drawRect.dx - textMetrics.width) * 0.5f + textMetrics.drawRect.x);
+        cacheSpriteOffset.x = (textMetrics.drawRect.dx - textMetrics.width) * 0.5f + textMetrics.drawRect.x;
     }
     else
     {
-        cacheSpriteOffset.x = (float32)textMetrics.drawRect.x;
+        cacheSpriteOffset.x = float32(textMetrics.drawRect.x);
     }
     if (align & ALIGN_BOTTOM)
     {
-        cacheSpriteOffset.y = (float32)(textMetrics.drawRect.dy - textMetrics.height + textMetrics.drawRect.y);
+        cacheSpriteOffset.y = float32(textMetrics.drawRect.dy - textMetrics.height + textMetrics.drawRect.y);
     }
     else if (align & ALIGN_VCENTER)
     {
-        cacheSpriteOffset.y = ((float32)(textMetrics.drawRect.dy - textMetrics.height) * 0.5f + textMetrics.drawRect.y);
+        cacheSpriteOffset.y = (textMetrics.drawRect.dy - textMetrics.height) * 0.5f + textMetrics.drawRect.y;
     }
     else
     {
-        cacheSpriteOffset.y = (float32)textMetrics.drawRect.y;
+        cacheSpriteOffset.y = float32(textMetrics.drawRect.y);
     }
 
     // Restore font size
